+++ src/test/java/wooteco/subway/domain/AgeTest.java
@@ -1,12 +1,11 @@
-package wooteco.subway;
+package wooteco.subway.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import wooteco.subway.domain.Age;
 
 class AgeTest {
